Move issue workflow from sonar-core to sonar-server
