+++ server/sonar-server/src/main/java/org/sonar/server/issue/workflow/FunctionExecutor.java
@@ -17,10 +17,9 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+package org.sonar.server.issue.workflow;
 
 import javax.annotation.Nullable;
-import org.sonar.api.batch.BatchSide;
 import org.sonar.api.issue.Issue;
 import org.sonar.api.server.ServerSide;
 import org.sonar.api.user.User;
@@ -28,7 +27,6 @@ import org.sonar.core.issue.DefaultIssue;
 import org.sonar.core.issue.IssueChangeContext;
 import org.sonar.core.issue.IssueUpdater;
 
-@BatchSide
 @ServerSide
 public class FunctionExecutor {
