+++ server/sonar-server/src/main/java/org/sonar/server/issue/workflow/IssueWorkflow.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+package org.sonar.server.issue.workflow;
 
 import java.util.List;
 import org.picocontainer.Startable;
