+++ server/sonar-server/src/main/java/org/sonar/server/issue/workflow/OrCondition.java
@@ -17,21 +17,26 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+package org.sonar.server.issue.workflow;
 
-import org.junit.Test;
+import org.sonar.api.issue.Issue;
+import org.sonar.api.issue.condition.Condition;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.sonar.core.issue.workflow.UnsetAssignee.INSTANCE;
+public class OrCondition implements Condition {
+  private final Condition[] conditions;
 
-public class UnsetAssigneeTest {
-
-  @Test
-  public void unassign() {
-    Function.Context context = mock(Function.Context.class);
-    INSTANCE.execute(context);
-    verify(context, times(1)).setAssignee(null);
+  public OrCondition(Condition... conditions) {
+    this.conditions = conditions;
   }
+
+  @Override
+  public boolean matches(Issue issue) {
+    for (Condition condition : conditions) {
+      if (condition.matches(issue)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
 }
