+++ server/sonar-server/src/main/java/org/sonar/server/issue/workflow/SetClosed.java
@@ -17,20 +17,26 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+package org.sonar.server.issue.workflow;
 
-import org.junit.Test;
+import org.sonar.api.issue.Issue;
+import org.sonar.core.issue.DefaultIssue;
 
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
+public enum SetClosed implements Function {
+  INSTANCE;
 
-public class SetResolutionTest {
-  @Test
-  public void execute() {
-    SetResolution function = new SetResolution("FIXED");
-    Function.Context context = mock(Function.Context.class);
-    function.execute(context);
-    verify(context, times(1)).setResolution("FIXED");
+  @Override
+  public void execute(Context context) {
+    DefaultIssue issue = (DefaultIssue) context.issue();
+    if (issue.isOnDisabledRule()) {
+      context.setResolution(Issue.RESOLUTION_REMOVED);
+    } else {
+      context.setResolution(Issue.RESOLUTION_FIXED);
+    }
+
+    // closed issues are not "tracked" -> the line number does not evolve anymore
+    // when code changes. That's misleading for end-users, so line number
+    // is unset.
+    context.setLine(null);
   }
 }
