+++ server/sonar-server/src/main/java/org/sonar/server/issue/workflow/State.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+package org.sonar.server.issue.workflow;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
