+++ server/sonar-server/src/main/java/org/sonar/server/issue/workflow/package-info.java
@@ -17,16 +17,8 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+@ParametersAreNonnullByDefault
+package org.sonar.server.issue.workflow;
 
-import org.sonar.api.issue.Issue;
-import org.sonar.api.issue.condition.Condition;
+import javax.annotation.ParametersAreNonnullByDefault;
 
-enum IsManual implements Condition {
-  INSTANCE;
-
-  @Override
-  public boolean matches(Issue issue) {
-    return issue.ruleKey().isManual();
-  }
-}
