+++ server/sonar-server/src/test/java/org/sonar/server/issue/workflow/SetResolutionTest.java
@@ -17,17 +17,20 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.issue.workflow;
+package org.sonar.server.issue.workflow;
 
-import org.sonar.api.issue.Issue;
-import org.sonar.api.issue.condition.Condition;
-import org.sonar.core.issue.DefaultIssue;
+import org.junit.Test;
 
-enum IsBeingClosed implements Condition {
-  INSTANCE;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
-  @Override
-  public boolean matches(Issue issue) {
-    return ((DefaultIssue) issue).isBeingClosed();
+public class SetResolutionTest {
+  @Test
+  public void execute() {
+    SetResolution function = new SetResolution("FIXED");
+    Function.Context context = mock(Function.Context.class);
+    function.execute(context);
+    verify(context, times(1)).setResolution("FIXED");
   }
 }
