+++ server/sonar-server/src/test/java/org/sonar/server/issue/ws/ShowActionTest.java
@@ -41,7 +41,7 @@ import org.sonar.core.issue.DefaultActionPlan;
 import org.sonar.core.issue.DefaultIssue;
 import org.sonar.core.issue.DefaultIssueComment;
 import org.sonar.core.issue.FieldDiffs;
-import org.sonar.core.issue.workflow.Transition;
+import org.sonar.server.issue.workflow.Transition;
 import org.sonar.core.user.DefaultUser;
 import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
