+++ sonar-batch/src/main/java/org/sonar/batch/scan/ProjectScanContainer.java
@@ -91,7 +91,6 @@ import org.sonar.batch.scan.measure.MeasureCache;
 import org.sonar.batch.source.CodeColorizers;
 import org.sonar.batch.test.TestPlanBuilder;
 import org.sonar.batch.test.TestableBuilder;
-import org.sonar.core.issue.workflow.FunctionExecutor;
 import org.sonar.core.metric.BatchMetrics;
 import org.sonar.core.platform.ComponentContainer;
 
@@ -175,7 +174,6 @@ public class ProjectScanContainer extends ComponentContainer {
       new QualityProfileProvider(),
 
       // issues
-      FunctionExecutor.class,
       IssueCache.class,
       DefaultProjectIssues.class,
       IssueTransition.class,
